requirements.txt
Pillow==9.1.0
easygui==0.98.3